+++ packages/Grit/src/package-test.ts
@@ -2,6 +2,7 @@ import { runTests, tests, TestContext , initAutoTests as initTests } from '@data
 import * as DG from 'datagrok-api/dg';
 
 import './tests/issue-crud-tests';
+import './tests/issue-labels-tests';
 
 export let _package = new DG.Package();
 export { tests };